requirements.txt
mlxtend==0.20.0
pandas==1.4.3
plotly==5.9.0
PyYAML==6.0
streamlit==1.11.0
streamlit_authenticator==0.2.1
